move tests into separate files in users/
